Move Jetpack migration ui to main.jetpack → .migration
