+++ WordPress/src/main/java/org/wordpress/android/ui/main/jetpack/migration/components/Message.kt
@@ -1,24 +1,24 @@
-package org.wordpress.android.ui.main.jetpack.components
+package org.wordpress.android.ui.main.jetpack.migration.components
 
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.unit.dp
-import org.wordpress.android.ui.compose.unit.FontSize
+import androidx.compose.ui.unit.sp
 
 @Composable
-fun Title(
+fun Message(
     text: String,
     modifier: Modifier = Modifier,
 ) {
     Text(
             text = text,
-            fontSize = FontSize.ExtraExtraExtraLarge.value,
-            fontWeight = FontWeight.Bold,
+            fontSize = 17.sp,
+            style = TextStyle(letterSpacing = (-0.01).sp),
             modifier = modifier
                     .padding(horizontal = 30.dp)
-                    .padding(top = 30.dp)
+                    .padding(top = 20.dp, bottom = 30.dp)
     )
 }
