+++ WordPress/src/main/java/org/wordpress/android/ui/main/jetpack/migration/components/SecondaryButton.kt
@@ -1,4 +1,4 @@
-package org.wordpress.android.ui.main.jetpack.components
+package org.wordpress.android.ui.main.jetpack.migration.components
 
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
@@ -7,25 +7,32 @@ import androidx.compose.material.ButtonDefaults
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
 
 @Composable
-fun PrimaryButton(
+fun SecondaryButton(
     text: String,
     onClick: () -> Unit,
-    modifier: Modifier = Modifier
+    modifier: Modifier = Modifier,
+    enabled: Boolean = true
 ) {
     Button(
             onClick = onClick,
+            enabled = enabled,
             elevation = ButtonDefaults.elevation(
                     defaultElevation = 0.dp,
                     pressedElevation = 0.dp,
             ),
+            colors = ButtonDefaults.buttonColors(
+                    backgroundColor = Color.Transparent
+            ),
             modifier = modifier
-                    .padding(top = 20.dp, bottom = 10.dp)
+                    .padding(bottom = 60.dp)
                     .padding(horizontal = 30.dp)
-                    .fillMaxWidth(),
+                    .fillMaxWidth()
     ) {
         Text(text = text)
     }
 }
+
