+++ WordPress/src/main/java/org/wordpress/android/ui/main/jetpack/migration/components/Subtitle.kt
@@ -1,24 +1,22 @@
-package org.wordpress.android.ui.main.jetpack.components
+package org.wordpress.android.ui.main.jetpack.migration.components
 
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
+import org.wordpress.android.ui.compose.unit.FontSize
 
 @Composable
-fun Message(
+fun Subtitle(
     text: String,
     modifier: Modifier = Modifier,
 ) {
     Text(
             text = text,
-            fontSize = 17.sp,
-            style = TextStyle(letterSpacing = (-0.01).sp),
+            fontSize = FontSize.ExtraLarge.value,
             modifier = modifier
                     .padding(horizontal = 30.dp)
-                    .padding(top = 20.dp, bottom = 30.dp)
+                    .padding(top = 20.dp)
     )
 }
