+++ WordPress/src/main/java/org/wordpress/android/ui/main/jetpack/migration/components/Title.kt
@@ -1,22 +1,24 @@
-package org.wordpress.android.ui.main.jetpack.components
+package org.wordpress.android.ui.main.jetpack.migration.components
 
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import org.wordpress.android.ui.compose.unit.FontSize
 
 @Composable
-fun Subtitle(
+fun Title(
     text: String,
     modifier: Modifier = Modifier,
 ) {
     Text(
             text = text,
-            fontSize = FontSize.ExtraLarge.value,
+            fontSize = FontSize.ExtraExtraExtraLarge.value,
+            fontWeight = FontWeight.Bold,
             modifier = modifier
                     .padding(horizontal = 30.dp)
-                    .padding(top = 20.dp)
+                    .padding(top = 30.dp)
     )
 }
